Add tests and documentation note for ams_pin
